[ fix ] common 모듈 수정으로 인한 페키지 및 코드 변경
