+++ movie-api/src/main/java/com/movie/movieapi/interfaces/movie/ReservationController.java
@@ -1,7 +1,7 @@
 package com.movie.movieapi.interfaces.movie;
 
 import com.movie.domain.facade.ReservationFacade;
-import com.movie.domain.response.Response;
+import com.movie.common.response.Response;
 import com.movie.movieapi.interfaces.movie.dto.ReservationDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpHeaders;
