+++ movie-api/src/test/java/com/movie/movieapi/interfaces/movie/ScheduleControllerTest.java
@@ -3,7 +3,7 @@ package com.movie.movieapi.interfaces.movie;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.movie.domain.movie.ScheduleService;
 import com.movie.domain.movie.dto.command.ScheduleCommand;
-import com.movie.domain.response.Response;
+import com.movie.common.response.Response;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
