+++ movie-domain/common/build.gradle
@@ -2,5 +2,5 @@ bootJar.enabled = false
 jar.enabled = true
 
 dependencies {
-
+    implementation 'org.springframework.boot:spring-boot-starter-web'
 }
