+++ movie-domain/common/src/main/java/com/movie/common/exception/ErrorCode.java
@@ -1,4 +1,4 @@
-package com.movie.domain.exception;
+package com.movie.common.exception;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
@@ -12,6 +12,7 @@ public enum ErrorCode {
     CONTENT_NOT_FOUND(HttpStatus.ACCEPTED, "CONTENT_NOT_FOUND"),
     LIMIT_EXCEEDED(HttpStatus.ACCEPTED, "LIMIT_EXCEEDED"),
     UNABLE_TO_RESERVE(HttpStatus.ACCEPTED, "UNABLE_TO_RESERVE"),
+    DISTRIBUTED_LOCK_NOT_AVAILABLE(HttpStatus.ACCEPTED,  "DISTRIBUTED_LOCK_NOT_AVAILABLE"),
 
 
     ;
