+++ movie-domain/common/src/main/java/com/movie/common/exception/GlobalControllerAdvice.java
@@ -1,6 +1,6 @@
-package com.movie.domain.exception;
+package com.movie.common.exception;
 
-import com.movie.domain.response.Response;
+import com.movie.common.response.Response;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
