+++ movie-domain/common/src/main/java/com/movie/common/response/Response.java
@@ -1,4 +1,4 @@
-package com.movie.domain.response;
+package com.movie.common.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
